Restore the discord round notification bot
