+++ Content.Server/_NF/RoundNotifications/Events/RoundStartedEvent.cs
@@ -1,0 +1,14 @@
+using Robust.Shared.Serialization;
+
+namespace Content.Server._NF.RoundNotifications.Events;
+
+[Serializable, NetSerializable]
+public sealed class RoundStartedEvent : EntityEventArgs
+{
+    public int RoundId { get; }
+
+    public RoundStartedEvent(int roundId)
+    {
+        RoundId = roundId;
+    }
+}
