--- Content.Shared/GameTicking/RoundRestartedEvent.cs
@@ -1,15 +1,0 @@
-﻿using Robust.Shared.Serialization;
-
-namespace Content.Shared.GameTicking
-{
-    [Serializable, NetSerializable]
-    public sealed class RoundStartedEvent : EntityEventArgs
-    {
-        public int RoundId { get; }
-
-        public RoundStartedEvent(int roundId)
-        {
-            RoundId = roundId;
-        }
-    }
-}
